+++ push_swap.h
@@ -6,7 +6,7 @@
 /*   By: lcalero <lcalero@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/09 17:27:04 by lcalero           #+#    #+#             */
-/*   Updated: 2025/01/14 17:09:00 by lcalero          ###   ########.fr       */
+/*   Updated: 2025/01/14 17:36:12 by lcalero          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,7 +60,7 @@ int		check_wrong_minus(int ac, char **av, int *stack_a, int single);
 
 void	ft_free(char **av, int size);
 
-size_t	countnum(char *argv);
+int		countnum(char *argv);
 
 //algorithm utils
 int		is_sorted(int *stack_a, int size_a);
